[poc/LEMS-3271-update-perseus-storybook] move back interactive graph stories to original location
